requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
idna==3.10
lxml==5.3.0
multidict==6.1.0
mysql-connector-python==9.1.0
outcome==1.3.0.post0
propcache==0.2.0
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.0
